refactor: link-level transition state instead of a pending banner
The pending signal moves into the Link itself: after a click, while the
router still shows the old page and the target's flight loads, the
clicked anchor carries data-pending/aria-busy and the outgoing page is
styled as stale (dimmed, cursor: progress) via body:has() in the global
stylesheet. The NavPending overlay (bar + chip) is removed — the
transition is an implementation detail of Link, not a screen element.
Pending resolves per link (clicked + location moved to this link's
target) and leaves the tree with the old page on swap; prefetched
targets swap too fast to ever show it.
