--- src/css/navPending.module.css
@@ -1,35 +1,0 @@
-.Bar {
-    position: fixed;
-    top: 0;
-    left: 0;
-    right: 0;
-    height: 3px;
-    background: linear-gradient(90deg, transparent, #f8d030, transparent);
-    background-size: 50% 100%;
-    background-repeat: no-repeat;
-    animation: slide 1s linear infinite;
-    z-index: 10;
-}
-
-@keyframes slide {
-    from { background-position: -100% 0; }
-    to { background-position: 200% 0; }
-}
-
-.Chip {
-    position: fixed;
-    bottom: 1rem;
-    right: 1rem;
-    background-color: #f8d030;
-    color: #000008;
-    font-weight: bold;
-    font-size: 13px;
-    padding: 0.4rem 0.8rem;
-    border-radius: 2px;
-    text-transform: capitalize;
-    z-index: 10;
-}
-
-@media (prefers-reduced-motion: reduce) {
-    .Bar { animation: none; }
-}
