--- src/css/navPending.module.css.d.ts
@@ -1,6 +1,0 @@
-declare const styles: {
-    Bar: string;
-    Chip: string;
-};
-
-export default styles
